chore: move network stuff into the network feature
