+++ src/features/common/data/data-store.ts
@@ -1,6 +1,6 @@
 import { updateClientConfig } from '@/features/common/data/algo-client'
 import { JotaiStore } from '@/features/common/data/types'
-import { NetworkConfigWithId } from '@/features/settings/data/types'
+import { NetworkConfigWithId } from '@/features/network/data/types'
 import { createStore } from 'jotai'
 import { useRef } from 'react'
