+++ src/features/layout/components/left-side-bar-menu.tsx
@@ -4,10 +4,11 @@ import { NavigationMenu, NavigationMenuItem, NavigationMenuLink, NavigationMenuL
 import { cn } from '@/features/common/utils'
 import { Button } from '@/features/common/components/button'
 import { useCallback } from 'react'
-import { useLayout, useSelectedNetwork } from '@/features/settings/data'
+import { useSelectedNetwork } from '@/features/network/data'
 import { Telescope, Braces, Settings, PanelLeftClose, PanelLeftOpen, ArrowLeft } from 'lucide-react'
 import { ThemeToggle } from '@/features/settings/components/theme-toggle'
 import { useNavigate } from 'react-router-dom'
+import { useLayout } from '@/features/settings/data'
 
 type Props = {
   className?: string
