+++ src/features/layout/pages/layout-page.test.tsx
@@ -5,8 +5,9 @@ import { LayoutPage } from '@/features/layout/pages/layout-page'
 import { connectWalletLabel, selectAccountLabel, disconnectWalletLabel } from '@/features/wallet/components/connect-wallet-button'
 import { PROVIDER_ID, Provider, useWallet } from '@txnlab/use-wallet'
 import { Event as TauriEvent, listen } from '@tauri-apps/api/event'
-import { networkConfigAtom, settingsStore } from '@/features/settings/data'
+import { networkConfigAtom } from '@/features/network/data'
 import { useNavigate } from 'react-router-dom'
+import { settingsStore } from '@/features/settings/data'
 
 describe('when rendering the layout page', () => {
   describe('and the wallet is not connected', () => {
