+++ src/features/network/components/create-network-config-form.tsx
@@ -2,16 +2,16 @@ import { Form } from '@/features/forms/components/form'
 import { useCallback, useMemo } from 'react'
 import { SubmitButton } from '@/features/forms/components/submit-button'
 import { FormActions } from '@/features/forms/components/form-actions'
-import { defaultNetworkConfigs, useNetworkConfigs, useSetCustomNetworkConfig } from '@/features/settings/data'
+import { defaultNetworkConfigs, useNetworkConfigs, useSetCustomNetworkConfig } from '@/features/network/data'
 import { z } from 'zod'
 import { toast } from 'react-toastify'
 import { CancelButton } from '@/features/forms/components/cancel-button'
-import { NetworkFormInner } from '@/features/settings/components/network-form-inner'
+import { NetworkFormInner } from '@/features/network/components/network-form-inner'
 import { asAlgoServiceConfig } from '@/features/settings/mappers'
 import { createNetworkConfigFormSchema } from '@/features/settings/form-schemas/create-network-config-form-schema'
 import { PROVIDER_ID } from '@txnlab/use-wallet'
 import { replaceAll } from '@/utils/replace-all.ts'
-import { tokenStorageText } from '@/features/settings/components/labels'
+import { tokenStorageText } from '@/features/network/components/labels'
 import { Alert } from '@/features/common/components/alert'
 
 type Props = {
