+++ src/features/network/components/networks-configs-table.tsx
@@ -1,15 +1,15 @@
 import { ColumnDef } from '@tanstack/react-table'
-import { defaultNetworkConfigs, useDeleteCustomNetworkConfig, useNetworkConfigs } from '@/features/settings/data'
+import { defaultNetworkConfigs, useDeleteCustomNetworkConfig, useNetworkConfigs } from '@/features/network/data'
 import { trimCharacterFromEnd } from '@/utils/trim-character-from-end'
 import { DataTable } from '@/features/common/components/data-table'
 import { Button } from '@/features/common/components/button'
 import { Dialog, DialogContent, DialogHeader, MediumSizeDialogBody } from '@/features/common/components/dialog'
 import { useCallback, useMemo, useState } from 'react'
-import { EditNetworkConfigForm } from '@/features/settings/components/edit-network-config-form'
-import { CreateNetworkConfigForm } from '@/features/settings/components/create-network-config-form'
+import { EditNetworkConfigForm } from '@/features/network/components/edit-network-config-form'
+import { CreateNetworkConfigForm } from '@/features/network/components/create-network-config-form'
 import { ConfirmButton } from '@/features/common/components/confirm-button'
 import { toast } from 'react-toastify'
-import { NetworkConfigWithId } from '@/features/settings/data/types'
+import { NetworkConfigWithId } from '@/features/network/data/types'
 import { Pencil, Plus, Trash } from 'lucide-react'
 
 export const networkConfigsTableLabel = 'Network Configs'
