+++ src/features/network/data/types.ts
@@ -1,0 +1,29 @@
+import { PROVIDER_ID } from '@txnlab/use-wallet'
+import { localnetId, mainnetId, testnetId } from '.'
+
+export type ServiceConfig = {
+  server: string
+  port: number
+  promptForToken?: boolean
+  token?: string
+}
+
+export type NetworkId = typeof localnetId | typeof testnetId | typeof mainnetId | string
+
+export type NetworkConfig = {
+  name: string
+  indexer: ServiceConfig
+  algod: ServiceConfig
+  kmd?: ServiceConfig
+  walletProviders: PROVIDER_ID[]
+}
+
+export type NetworkConfigWithId = {
+  id: NetworkId
+} & NetworkConfig
+
+export type NetworkTokens = {
+  algod?: string
+  indexer?: string
+  kmd?: string
+}
