+++ src/features/settings/data/index.ts
@@ -1,5 +1,4 @@
 export * from './settings'
-export * from './network'
 export * from './theme'
 export * from './layout'
 export * from './types'
