+++ src/index-page.tsx
@@ -1,8 +1,8 @@
-import { useSelectedNetwork } from '@/features/settings/data'
 import { useNavigate } from 'react-router-dom'
 import { Urls } from '@/routes/urls'
 import { Loader } from 'lucide-react'
 import { useEffect } from 'react'
+import { useSelectedNetwork } from './features/network/data'
 
 export function IndexPage() {
   const [selectedNetwork] = useSelectedNetwork()
